requirements.txt
discord.py>=2.3.0
python-dotenv
openai
aiohttp
playwright
beautifulsoup4
youtube-transcript-api
requests
torch
openai-whisper
pydub
